+++ lib/mui/key_handler/normal_mode.rb
@@ -500,9 +500,9 @@ module Mui
         return result(message: "No previous search pattern") unless @search_state&.has_pattern?
 
         match = if @search_state.direction == :forward
-                  @search_state.find_next(cursor_row, cursor_col)
+                  @search_state.find_next(cursor_row, cursor_col, buffer:)
                 else
-                  @search_state.find_previous(cursor_row, cursor_col)
+                  @search_state.find_previous(cursor_row, cursor_col, buffer:)
                 end
 
         if match
@@ -517,9 +517,9 @@ module Mui
         return result(message: "No previous search pattern") unless @search_state&.has_pattern?
 
         match = if @search_state.direction == :forward
-                  @search_state.find_previous(cursor_row, cursor_col)
+                  @search_state.find_previous(cursor_row, cursor_col, buffer:)
                 else
-                  @search_state.find_next(cursor_row, cursor_col)
+                  @search_state.find_next(cursor_row, cursor_col, buffer:)
                 end
 
         if match
@@ -538,13 +538,12 @@ module Mui
         escaped_pattern = "\\b#{Regexp.escape(word)}\\b"
 
         @search_state.set_pattern(escaped_pattern, direction)
-        @search_state.find_all_matches(buffer)
 
         # Find next/previous match from current position
         match = if direction == :forward
-                  @search_state.find_next(cursor_row, cursor_col)
+                  @search_state.find_next(cursor_row, cursor_col, buffer:)
                 else
-                  @search_state.find_previous(cursor_row, cursor_col)
+                  @search_state.find_previous(cursor_row, cursor_col, buffer:)
                 end
 
         if match